+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - Use new label builders ([#454]).
 - Change the liveness probes to use the web UI port and to fail after
   one minute ([#491]).
+- Update the getting started script to use the listener class cluster-internal ([#492])
 
 ### Removed
 
@@ -40,6 +41,7 @@ All notable changes to this project will be documented in this file.
 [#474]: https://github.com/stackabletech/hdfs-operator/pull/474
 [#475]: https://github.com/stackabletech/hdfs-operator/pull/475
 [#491]: https://github.com/stackabletech/hdfs-operator/pull/491
+[#492]: https://github.com/stackabletech/hdfs-operator/pull/492
 
 ## [23.11.0] - 2023-11-24